Update setup.py
The setup.py file only checks for 1 .dll file out of the many available and it caused me a great deal of pain. 

Either fix this, or don't give the option to only install models that you need when installing the PicoSDK. It said 'which libraries would you like to install' and obviously I only installed the ones for the model that I had (ps4000).

I have discussed this over the phone a month ago with one of your colleagues and they said it would be fixed, which it hasn't.